LayoutTests/ipc/isasyncreply-exposed-to-ipc-testing-api.html

<!doctype html><!-- webkit-test-runner [ IPCTestingAPIEnabled=true ] -->
<title>Test that isAsyncReply is correctly exposed to the IPCTestingAPI</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/ipc.js"></script>
<body>
<p>Test pass if assertions don't cause error</p>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

if (window.IPC) {
    assert_equals(IPC.messages.IPCTester_AsyncOptionalExceptionData.isAsyncReply, false);
    assert_equals(IPC.messages.IPCTester_AsyncOptionalExceptionDataReply.isAsyncReply, true);
    assert_equals(IPC.messages.WebPageTesting_IsLayerTreeFrozen.isAsyncReply, false);
    assert_equals(IPC.messages.WebPageTesting_IsLayerTreeFrozenReply.isAsyncReply, true);
    assert_equals(IPC.messages.WebPageProxy_DecidePolicyForNavigationActionAsync.isAsyncReply, false);
    assert_equals(IPC.messages.WebPageProxy_DecidePolicyForNavigationActionAsyncReply.isAsyncReply, true);
    assert_equals(IPC.messages.RemoteMediaResourceManager_ResponseReceived.isAsyncReply, false);
    assert_equals(IPC.messages.RemoteMediaResourceManager_ResponseReceivedReply.isAsyncReply, true);
}

document.write("PASS");
testRunner.notifyDone();
</script>
</body>
